ct: move extent_meta into level zero
it is currently only used in level zero. its object id also embeds an
epoch, making it even more unique to level zero. a similar structure
will likely be created in the future which is unified across all levels,
and will most likely be related to unified caching operations.
