+++ src/v/cloud_topics/batch_cache/BUILD
@@ -18,9 +18,9 @@ redpanda_cc_library(
         "//src/v/base",
         "//src/v/bytes",
         "//src/v/bytes:iobuf",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/config",
         "//src/v/model",
