+++ src/v/cloud_topics/data_plane_api.h
@@ -11,7 +11,7 @@
 #pragma once
 
 #include "base/outcome.h"
-#include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/common/extent_meta.h"
 #include "container/chunked_circular_buffer.h"
 #include "container/fragmented_vector.h"
 #include "model/fundamental.h"
