+++ src/v/cloud_topics/level_zero/batcher/tests/BUILD
@@ -36,10 +36,10 @@ redpanda_cc_gtest(
         "//src/v/cloud_io:io_result",
         "//src/v/cloud_io:remote",
         "//src/v/cloud_storage_clients",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:object_utils",
         "//src/v/cloud_topics/level_zero/batcher",
         "//src/v/cloud_topics/level_zero/batcher:aggregator",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/cloud_topics/level_zero/pipeline:event_filter",
         "//src/v/cloud_topics/level_zero/pipeline:serializer",
         "//src/v/cloud_topics/level_zero/pipeline:write_pipeline",
