+++ src/v/cloud_topics/level_zero/batcher/tests/batcher_test.cc
@@ -13,8 +13,8 @@
 #include "bytes/iostream.h"
 #include "cloud_io/io_result.h"
 #include "cloud_io/remote.h"
-#include "cloud_topics/extent_meta.h"
 #include "cloud_topics/level_zero/batcher/batcher.h"
+#include "cloud_topics/level_zero/common/extent_meta.h"
 #include "cloud_topics/object_utils.h"
 #include "model/fundamental.h"
 #include "model/namespace.h"
