+++ src/v/cloud_topics/level_zero/common/extent_meta.h
