+++ src/v/cloud_topics/level_zero/pipeline/BUILD
@@ -26,7 +26,7 @@ redpanda_cc_library(
     deps = [
         "//src/v/base",
         "//src/v/bytes:iobuf",
-        "//src/v/cloud_topics:extent_meta",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
     ],
@@ -75,8 +75,8 @@ redpanda_cc_library(
         ":pipeline_stage",
         ":serializer",
         "//src/v/base",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/container:fragmented_vector",
         "//src/v/container:intrusive",
         "//src/v/model",
@@ -100,8 +100,8 @@ redpanda_cc_library(
         ":pipeline_stage",
         ":serializer",
         "//src/v/base",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/container:fragmented_vector",
         "//src/v/container:intrusive",
         "//src/v/model",
@@ -202,8 +202,8 @@ redpanda_cc_library(
         "//src/v/base",
         "//src/v/bytes",
         "//src/v/bytes:iobuf",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/config",
         "//src/v/model",
         "//src/v/ssx:semaphore",
