+++ src/v/cloud_topics/level_zero/pipeline/serializer.cc
@@ -10,7 +10,7 @@
 
 #include "cloud_topics/level_zero/serializer.h"
 
-#include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/common/extent_meta.h"
 #include "model/timeout_clock.h"
 #include "storage/record_batch_utils.h"
