+++ src/v/cloud_topics/level_zero/pipeline/serializer.h
@@ -12,7 +12,7 @@
 
 #include "base/seastarx.h"
 #include "bytes/iobuf.h"
-#include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/common/extent_meta.h"
 #include "container/fragmented_vector.h"
 #include "model/record_batch_reader.h"
