+++ src/v/cloud_topics/level_zero/pipeline/tests/event_filter_test.cc
@@ -8,7 +8,7 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/common/extent_meta.h"
 #include "cloud_topics/level_zero/event_filter.h"
 #include "cloud_topics/level_zero/write_pipeline.h"
 #include "container/fragmented_vector.h"
