+++ src/v/cloud_topics/level_zero/pipeline/write_request.h
@@ -13,7 +13,7 @@
 #include "base/outcome.h"
 #include "base/seastarx.h"
 #include "cloud_topics/errc.h"
-#include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/common/extent_meta.h"
 #include "cloud_topics/level_zero/pipeline_stage.h"
 #include "cloud_topics/level_zero/serializer.h"
 #include "container/fragmented_vector.h"
