+++ src/v/cloud_topics/level_zero/reader/BUILD
@@ -23,10 +23,10 @@ redpanda_cc_library(
         "//src/v/cloud_io:basic_cache_service_api",
         "//src/v/cloud_io:io_result",
         "//src/v/cloud_io:remote",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:object_utils",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/config",
         "//src/v/model",
@@ -55,9 +55,9 @@ redpanda_cc_library(
         "//src/v/cloud_io:basic_cache_service_api",
         "//src/v/cloud_io:io_result",
         "//src/v/cloud_io:remote",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:logger",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/cloud_topics/level_zero/reader:materialized_extent",
         "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/config",
@@ -119,7 +119,7 @@ redpanda_cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//src/v/base",
-        "//src/v/cloud_topics:extent_meta",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/model",
         "//src/v/storage",
         "@boost//:outcome",
