+++ src/v/cloud_topics/level_zero/reader/reader.h
@@ -9,7 +9,7 @@
  */
 #pragma once
 
-#include "cloud_topics/extent_meta.h"
+#include "cloud_topics/level_zero/common/extent_meta.h"
 #include "model/record_batch_reader.h"
 #include "storage/log_reader.h"
