+++ src/v/cloud_topics/level_zero/reader/tests/BUILD
@@ -20,9 +20,9 @@ redpanda_test_cc_library(
         "//src/v/cloud_io:io_result",
         "//src/v/cloud_io:remote",
         "//src/v/cloud_storage_clients",
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:object_utils",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/cloud_topics/level_zero/reader:materialized_extent",
         "//src/v/cloud_topics/level_zero/stm:placeholder",
         "//src/v/container:fragmented_vector",
@@ -46,8 +46,8 @@ redpanda_cc_gtest(
         "materialized_extent_test.cc",
     ],
     deps = [
-        "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:types",
+        "//src/v/cloud_topics/level_zero/common:extent_meta",
         "//src/v/cloud_topics/level_zero/reader:materialized_extent",
         "//src/v/cloud_topics/level_zero/reader/tests:materialized_extent_fixture",
         "//src/v/cloud_topics/level_zero/stm:placeholder",
